MSVC: enable C4388/C4389 (signed/unsigned equality) and fix the sites
Mirror the gcc -Wsign-compare ratchet on the cl.exe side: enable C4388/C4389
(equality signed/unsigned mismatch, off by default) via DmsDef.props
AdditionalOptions (/w14388 /w14389). C4018 (relational) was already clean.

These 8 equality sites were flagged by MSVC but not gcc, in two groups:
- genuine LLP64-vs-LP64 long-width divergence: dbfImp.cpp:283/297
  `ftell()` (long: 32-bit on Win, 64-bit on Linux) == ActualPosition() (UInt32).
  On gcc the UInt32 promotes into 64-bit signed long (no warning); on MSVC both
  go unsigned. Fixed as UInt32(ftell(...)) == ActualPosition(...) (dbf positions
  are 32-bit by design; behaviour identical incl. the ftell()==-1 error case).
- Windows-only code gcc never compiles (so it only ever warned on MSVC):
  FileMapHandle.cpp:96/484/501/510 (#if WIN32 mmap impl) and ViewPort.cpp:937/942
  (#ifdef _WIN32 SaveBitmap tile loop). Sentinel `== -1` made typed, `1<<y` ->
  `UInt32(1)<<y`, and the UInt32 loop indices compared against the non-negative
  IPoint tile counts via explicit UInt32(...) (already the implicit conversion).

All behaviour-preserving. (Permission to edit the shared DmsDef.props was granted.)

+++ stg/dll/src/dbf/dbfImp.cpp
@@ -280,7 +280,7 @@ bool DbfImpl::GoTo(UInt32 recNo)
 
 bool DbfImpl::IsAt(UInt32 recNo) const
 {
-	return ftell(GetFP()) == ActualPosition(recNo);
+	return UInt32(ftell(GetFP())) == ActualPosition(recNo);
 }
 
 bool DbfImpl::ReadRecord  (void* buffer)
@@ -294,7 +294,7 @@ bool DbfImpl::AppendRecord(const void* buffer)
 {
 	dms_assert(GetFP());
 
-	dms_assert(ftell(GetFP()) == ActualPosition(m_RecordCount));
+	dms_assert(UInt32(ftell(GetFP())) == ActualPosition(m_RecordCount));
 	if (fwrite(buffer, m_RecordSize, 1, GetFP()) != 1)
 		return false;
 	++m_RecordCount;
